Refactor templateBars
